update: Added filter data system. * Modified MunroModel to use Double with Default items. * NB: must modify HillCategory to potentially use data HillcategoryType.

+++ app/src/main/java/com/katocoding/munrolibrarychallenge/data/munrolist/MunroListExtractor.kt
@@ -21,7 +21,7 @@ class MunroListExtractor {
         return records.map {
             MunroModel(
                 name = it[namePos],
-                heightM = it[heightMPos],
+                heightM = it[heightMPos].toDouble(),
                 gridReference = it[gridRefPos],
                 hillCategory = it[hillCategoryPos]
             )
